change package name to new format hudson.plugins, add Extension and DataBoundConstructor annotation to be compatible with new jenkins, added 1 unit test, more to come
