+++ src/main/resources/hudson/plugins/s3/S3BucketPublisher/config.jelly
